==> scripts/genotype/06-THRIVE_VR-pi_comparisons-plot.R <==
library(tidyverse)
library(here)
pisComp <- read_csv(here("genomics/pixy-pi", "240617_samples_pi.csv"))
pisComp_us <- subset(pisComp, us == "1")
pisComp_alb <- subset(pisComp, species == "albicans")
pisComp_alb_11 <- subset(pisComp_alb, number_of_samples == 11)
pisComp_alb <- subset(pisComp_alb, number_of_samples < 5)

pdf(here("genomics", "figures", "240707_C.alb_study_comparisons.pdf"), width = 5, height = 7)
plot(pisComp_alb$place, pisComp_alb$average_diversity, ylim=c(0.002, 0.006), yaxt="n", pch = pisComp_alb$pch, ylab="", xlab="", xaxt="n", cex=1.25)
#points(pisComp_alb_11$place, pisComp_alb_11$average_diversity, pch = pisComp_alb_11$pch, cex = 1.25)
axis(2, las=2)
axis(1, at=1:12, c("TVY4", "TYV10", "YST7", "814", "833", "811", "813", "830", "882","A", "D", "G"), cex.axis=0.8, )
abline(v=3.5, lty=2)
abline(v=9.5, lty=2)
legend("topleft", c("vaginal", "rectal", "oral"), pch=c(0, 2, 1), inset = 0.02)
dev.off()
pisComp_glab <- subset(pisComp, species == "glabrata")
pisComp_glab <- subset(pisComp_glab, number_of_samples > 5)

pdf(here("genomics", "figures", "240707_N.glab_study_comparisons.pdf"), width = 3.1, height = 3.9)
plot(pisComp_glab$place, pisComp_glab$average_diversity, yaxt="n", pch = pisComp_glab$pch, ylab="", xlab="", xaxt="n", cex=1.25)
axis(2, las=2)
axis(1, at=1:10, c("YST6", pisComp_glab$pop[1:9]), cex.axis=0.8, )
abline(v=1.5, lty=2)
legend("bottomright", c("vaginal", "rectal", "blood"), pch=c(0, 2, 5), inset = 0.02)
dev.off()
